backend/requirements.txt
mysql-connector-python
dotenv
flask
flask-cors